Refactors battle scene resource loading
Consolidates battle scene resource loading into a single function for better management.

This change improves code organization and reduces redundancy by
centralizing the logic for loading and merging resources required for
battle scenes. It replaces the previous separate loading and merging steps
with a single `updateBattleSceneResources` function.
