+++ src/js/game/game-procedure/start-episode.ts
@@ -2,7 +2,7 @@ import { fadeOut, stop } from "../../bgm/bgm-operators";
 import { MAX_LOADING_TIME } from "../../dom-scenes/dom-scene-binder/max-loading-time";
 import { EpisodeTitle } from "../../dom-scenes/episode-title";
 import { NPCBattleRoom } from "../../npc/npc-battle-room";
-import { loadAdditionalBattleSceneResources } from "../../resource/loading/load-additional-battle-scene-resources";
+import { updateBattleSceneResources } from "../../resource/update-battle-scene-resources";
 import { mergeResources } from "../../resource/loading/merge-resources";
 import { BattleScene } from "../../td-scenes/battle";
 import { waitAnimationFrame } from "../../wait/wait-animation-frame";
@@ -42,17 +42,10 @@ export async function startEpisode(
   const startTutorialStageTime = Date.now();
   const config = await props.config.load();
   props.renderer.setPixelRatio(config.webGLPixelRatio);
-
-  const additionalLoading = loadAdditionalBattleSceneResources({
-    ...props,
+  props.resources = await updateBattleSceneResources({
+    resources: props.resources,
     players: [npcBattle.player, npcBattle.enemy],
   });
-  const additionalResources = await additionalLoading.resources;
-  props.resources = mergeResources({
-    resources: props.resources,
-    loaded: additionalResources,
-  });
-
   const battleScene = new BattleScene({
     ...props,
     playingBGM: episode.bgm,
